Migration AGILE-> Eclipse AGAIL (#5)
* Update License Header

* Update dbus params in test scripts

* Migration AGILE -> Eclipse AGAIL

* Update README.md
