+++ iot.agail.protocol.iotivity/src/common/DelayedCallback.cpp
@@ -1,31 +1,21 @@
-/*
- * Copyright 2016 CREATE-NET
+/*******************************************************************************
+ * Copyright (c) 2016, 2017 FBK CREATE-NET
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * and Eclipse Distribution License v1.0 which accompany this distribution.
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ * The Eclipse Public License is available at
+ *    http://www.eclipse.org/legal/epl-v10.html
+ * and the Eclipse Distribution License is available at
+ *   http://www.eclipse.org/org/documents/edl-v10.php.
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * @ingroup
- * @{
+ * Description: DelayedCallback
+ *              DelayedCallback. Implementation file
  *
- * @file        DelayedCallback.cpp
- * @brief       
- *
- * @author      Mattia Antonini <mattia.antonini1@studenti.unipr.it>
- *                              <m.antonini@create-net.org>
- *
- * @}
- */
+ * Contributors:
+ *    Mattia Antonini
+ *******************************************************************************/
 
 #include "DelayedCallback.h"
