--- iot.agile.protocol.iotivity/src/Resource.cpp
@@ -1,50 +1,0 @@
-/*
- * Copyright 2016 CREATE-NET
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * @ingroup
- * @{
- *
- * @file        Resource.cpp
- * @brief       Implementation file for Resource wrapper
- *
- * @author      Mattia Antonini <mattia.antonini1@studenti.unipr.it>
- *                              <m.antonini@create-net.org>
- *
- * @}
- */
-
-#include "Resource.h"
-
-Resource::Resource()
-{
-    resource = nullptr;
-}
-
-Resource::Resource(std::shared_ptr<OC::OCResource> res)
-{
-    resource = res;
-}
-
-Resource::~Resource()
-{
-}
-
-bool Resource::operator==(const Resource &res) const
-{
-    const Resource t = *this;
-    return (*(t.resource) == *(res.resource));
-}
