--- iot.agile.protocol.iotivity/src/Resource.h
@@ -1,45 +1,0 @@
-/*
- * Copyright 2016 CREATE-NET
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * @ingroup
- * @{
- *
- * @file        Resource.h
- * @brief       Header file for Resource wrapper
- *
- * @author      Mattia Antonini <mattia.antonini1@studenti.unipr.it>
- *                              <m.antonini@create-net.org>
- *
- * @}
- */
-
-#include <unistd.h>
-
-#include <iotivity/resource/OCResource.h>
-
-
-class Resource
-{
-    public:
-    std::shared_ptr<OC::OCResource> resource;
-
-    Resource();
-    Resource(std::shared_ptr<OC::OCResource>);
-    ~Resource();
-
-    bool operator==(const Resource&) const;
-};
