--- iot.agile.protocol.iotivity/src/agile/DBusUtility.h
@@ -1,38 +1,0 @@
-/*
- * Copyright 2016 CREATE-NET
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * @ingroup
- * @{
- *
- * @file        DBusUtility.h
- * @brief       Header file for DBusUtility
- *
- * @author      Mattia Antonini <mattia.antonini1@studenti.unipr.it>
- *                              <m.antonini@create-net.org>
- *
- * @}
- */
-
-#ifndef __DBUS_UTILITY_H__
-#define __DBUS_UTILITY_H__
-
-#include <gio/gunixfdlist.h>
-#include <gio/gio.h>
-
-GVariant* g_variant_deep_copy(GVariant *);
-
-#endif
