--- iot.agile.protocol.iotivity/src/common/utility.cpp
@@ -1,48 +1,0 @@
-/*
- * Copyright 2016 CREATE-NET
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * @ingroup
- * @{
- *
- * @file        utility.cpp
- * @brief       Utility functions
- *
- * @author      Mattia Antonini <mattia.antonini1@studenti.unipr.it>
- *                              <m.antonini@create-net.org>
- *
- * @}
- */
-
-#include "utility.h"
-
-bool is_ipv6_address(const string& str)
-{
-#if PRINT_PRETTY_LOGS
-    cerr << "Function: " << __PRETTY_FUNCTION__ << std::endl;
-#endif
-    struct sockaddr_in6 sa;
-    return inet_pton(AF_INET6, str.c_str(), &(sa.sin6_addr))!=0;
-}
-
-bool is_ipv4_address(const string& str)
-{
-#if PRINT_PRETTY_LOGS
-    cerr << "Function: " << __PRETTY_FUNCTION__ << std::endl;
-#endif
-    struct sockaddr_in sa;
-    return inet_pton(AF_INET, str.c_str(), &(sa.sin_addr))!=0;
-}
