--- iot.agile.protocol.iotivity/src/common/utility.h
@@ -1,51 +1,0 @@
-/*
- * Copyright 2016 CREATE-NET
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * @ingroup
- * @{
- *
- * @file        utility.h
- * @brief       Header file for Utility functions
- *
- * @author      Mattia Antonini <mattia.antonini1@studenti.unipr.it>
- *                              <m.antonini@create-net.org>
- *
- * @}
- */
-
-#ifndef __COMMON_UTILITY_H__
-#define __COMMON_UTILITY_H__
-
-#include <unistd.h>
-#include <ctime>
-#include <string>
-#include <algorithm>
-#include <typeinfo>
-#include <iostream>
-#include <vector>
-#include <termios.h>
-
-#include <sys/socket.h>
-#include <netinet/in.h>
-#include <arpa/inet.h>
-
-using namespace std;
-
-bool is_ipv6_address(const string&);
-bool is_ipv4_address(const string& str);
-
-#endif
