+++ scripts/stop.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 ARG="$1"
-PATTERN="iot.agile.$ARG"
+PATTERN="iot.agail.$ARG"
 
 echo "Send kill signal to $PATTERN"
 pkill -f $PATTERN
